requirements.txt
pytorch-lightning>=1.0.2
torch>=1.0.2
torchvision>=0.17.2
numpy~=1.26.4
nibabel~=5.2.1
scipy~=1.12.0
setuptools~=69.0.3
pandas~=2.2.1
volumentations-3D~=1.0.4
wandb~=0.17.0
